requirements.txt
streamlit-lottie==0.0.5
plotly==5.20.0
pygwalker==0.4.7
scikit-learn==1.4.2